refactor(calculator): CalculatorChunkProcessingCoordinator uses common ObjectStorage
Migrates from maple.calculator.storage.ObjectStorage to
maple.expectation.common.storage.ObjectStorage (VS1).

The coordinator only uses exists() which has the same signature in
both types. Mock target rename in CalculatorChunkProcessingCoordinatorTest.

Default backend = local. No behavior change (LocalFsObjectStorage
behaves identically to LocalObjectStorageAdapter for this path).

Issue #1217
