+++ module-calculator/src/main/kotlin/maple/calculator/CalculatorChunkProcessingCoordinator.kt
@@ -10,7 +10,7 @@ import maple.calculator.event.KafkaResultEventPublisher
 import maple.calculator.metrics.CalculatorMetricsListener
 import maple.calculator.model.ChunkResult
 import maple.calculator.processor.SnapshotChunkProcessor
-import maple.calculator.storage.ObjectStorage
+import maple.expectation.common.storage.ObjectStorage
 import maple.expectation.common.event.CalculatorResultChunkReadyEvent
 import maple.expectation.common.event.SnapshotChunkReadyEvent
 import maple.expectation.util.CompressionUtils
